Fix Index-64 conversion errors

+++ TESTING/EIG/cchkbd.f
@@ -562,7 +562,7 @@
          M = MVAL( JSIZE )
          N = NVAL( JSIZE )
          MNMIN = MIN( M, N )
-         AMNINV = ONE / MAX( M, N, 1 )
+         AMNINV = ONE / REAL( MAX( M, N, 1 ) )
 *
          IF( NSIZES.NE.1 ) THEN
             MTYPES = MIN( MAXTYP, NTYPES )
@@ -619,7 +619,7 @@
             GO TO 70
 *
    60       CONTINUE
-            ANORM = RTUNFL*MAX( M, N )*ULPINV
+            ANORM = RTUNFL*REAL( MAX( M, N ) )*ULPINV
             GO TO 70
 *
    70       CONTINUE
