+++ TESTING/LIN/cqrt05.f
@@ -182,7 +182,7 @@
       ANORM = CLANGE( '1', M2, N, A, M2, RWORK )
       RESID = CLANGE( '1', M2, N, R, M2, RWORK )
       IF( ANORM.GT.ZERO ) THEN
-         RESULT( 1 ) = RESID / (EPS*ANORM*MAX(1,M2))
+         RESULT( 1 ) = RESID / (EPS*ANORM*REAL( MAX( 1, M2 ) ))
       ELSE
          RESULT( 1 ) = ZERO
       END IF
@@ -193,7 +193,7 @@
       CALL CHERK( 'U', 'C', M2, M2, REAL(-ONE), Q, M2, REAL(ONE),
      $            R, M2 )
       RESID = CLANSY( '1', 'Upper', M2, R, M2, RWORK )
-      RESULT( 2 ) = RESID / (EPS*MAX(1,M2))
+      RESULT( 2 ) = RESID / (EPS*REAL( MAX( 1, M2 ) ))
 *
 *     Generate random m-by-n matrix C and a copy CF
 *
@@ -213,7 +213,7 @@
       CALL CGEMM( 'N', 'N', M2, N, M2, -ONE, Q, M2, C, M2, ONE, CF, M2 )
       RESID = CLANGE( '1', M2, N, CF, M2, RWORK )
       IF( CNORM.GT.ZERO ) THEN
-         RESULT( 3 ) = RESID / (EPS*MAX(1,M2)*CNORM)
+         RESULT( 3 ) = RESID / (EPS*REAL( MAX( 1, M2 ) )*CNORM)
       ELSE
          RESULT( 3 ) = ZERO
       END IF
@@ -232,7 +232,7 @@
       CALL CGEMM('C','N',M2,N,M2,-ONE,Q,M2,C,M2,ONE,CF,M2)
       RESID = CLANGE( '1', M2, N, CF, M2, RWORK )
       IF( CNORM.GT.ZERO ) THEN
-         RESULT( 4 ) = RESID / (EPS*MAX(1,M2)*CNORM)
+         RESULT( 4 ) = RESID / (EPS*REAL( MAX( 1, M2 ) )*CNORM)
       ELSE
          RESULT( 4 ) = ZERO
       END IF
@@ -255,7 +255,7 @@
       CALL CGEMM('N','N',N,M2,M2,-ONE,D,N,Q,M2,ONE,DF,N)
       RESID = CLANGE('1',N, M2,DF,N,RWORK )
       IF( CNORM.GT.ZERO ) THEN
-         RESULT( 5 ) = RESID / (EPS*MAX(1,M2)*DNORM)
+         RESULT( 5 ) = RESID / (EPS*REAL( MAX( 1, M2 ) )*DNORM)
       ELSE
          RESULT( 5 ) = ZERO
       END IF
@@ -275,7 +275,7 @@
       CALL CGEMM( 'N', 'C', N, M2, M2, -ONE, D, N, Q, M2, ONE, DF, N )
       RESID = CLANGE( '1', N, M2, DF, N, RWORK )
       IF( CNORM.GT.ZERO ) THEN
-         RESULT( 6 ) = RESID / (EPS*MAX(1,M2)*DNORM)
+         RESULT( 6 ) = RESID / (EPS*REAL( MAX( 1, M2 ) )*DNORM)
       ELSE
          RESULT( 6 ) = ZERO
       END IF
@@ -285,4 +285,3 @@
       DEALLOCATE ( A, AF, Q, R, RWORK, WORK, T, C, D, CF, DF)
       RETURN
       END
-
